# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface SeatInfo {
  stadiumId: string;
  section: string;
  row: string;
  seat: string;
}

export interface ScanTicketBody {
  qrCode: string;
}

export interface FanSession {
  id: string;
  stadiumId: string;
  stadiumName: string;
  section: string;
  row: string;
  seat: string;
  fanName: string;
  goalPoints: number;
  ordersCount: number;
  createdAt: string;
}

export interface Stadium {
  id: string;
  name: string;
  city: string;
  active: boolean;
  sections: string[];
}

export type StandStatus = (typeof StandStatus)[keyof typeof StandStatus];

export const StandStatus = {
  open: "open",
  busy: "busy",
  temp_closed: "temp_closed",
  closed: "closed",
} as const;

export interface Stand {
  id: string;
  stadiumId: string;
  name: string;
  description?: string;
  category: string;
  status: StandStatus;
  avgPrepTimeMinutes: number;
  pendingOrders: number;
  maxPendingOrders: number;
  waitTimeMinutes: number;
  imageUrl: string;
  rating: number;
  location?: string;
  reopenAt?: string;
  isHalalOnly?: boolean;
}

export interface MenuItem {
  id: string;
  standId: string;
  name: string;
  description: string;
  priceMad: number;
  category: string;
  allergens: string[];
  isHalal: boolean;
  isAvailable: boolean;
  imageUrl: string;
  isVegetarian?: boolean;
  isFeatured?: boolean;
}

export interface OrderItemInput {
  menuItemId: string;
  /** @minimum 1 */
  quantity: number;
}

export interface OrderItem {
  menuItemId: string;
  name: string;
  quantity: number;
  priceMad: number;
  imageUrl?: string;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  received: "received",
  preparing: "preparing",
  on_the_way: "on_the_way",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: string;
  sessionId: string;
  standId: string;
  standName: string;
  riderId?: string;
  riderName?: string;
  items: OrderItem[];
  status: OrderStatus;
  seat: SeatInfo;
  subtotalMad: number;
  deliveryFeeMad: number;
  totalMad: number;
  etaMinutes: number;
  createdAt: string;
  scheduledFor?: string;
  deliveredAt?: string;
  rating?: number;
  isPreOrder: boolean;
  fanName?: string;
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  card: "card",
  apple_pay: "apple_pay",
  google_pay: "google_pay",
} as const;

export interface CreateOrderBody {
  sessionId: string;
  standId: string;
  items: OrderItemInput[];
  scheduledFor?: string;
  paymentMethod?: CreateOrderBodyPaymentMethod;
}

export interface UpdateOrderStatusBody {
  status: OrderStatus;
  riderId?: string;
}

export interface RateOrderBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
}

export type RiderStatus = (typeof RiderStatus)[keyof typeof RiderStatus];

export const RiderStatus = {
  available: "available",
  busy: "busy",
  offline: "offline",
} as const;

export interface Rider {
  id: string;
  name: string;
  stadiumId: string;
  status: RiderStatus;
  activeOrders: number;
  avatarUrl?: string;
}

export interface RiderLoginBody {
  pin: string;
}

export interface RiderStats {
  riderId: string;
  deliveredToday: number;
  activeOrders: number;
  avgDeliveryMinutes: number;
}

export interface Admin {
  id: string;
  email: string;
  stadiumId: string;
  stadiumName: string;
}

export interface AdminLoginBody {
  email: string;
  password: string;
}

export type AdminDashboardOrdersByHourItem = {
  hour: string;
  count: number;
};

export type AdminDashboardTopItemsItem = {
  name: string;
  count: number;
  standName: string;
};

export type AdminDashboardOrdersByStandItem = {
  standName: string;
  count: number;
  revenueMad: number;
};

export interface AdminDashboard {
  ordersToday: number;
  revenueTodayMad: number;
  activeRiders: number;
  avgDeliveryMinutes: number;
  ordersByHour: AdminDashboardOrdersByHourItem[];
  topItems: AdminDashboardTopItemsItem[];
  ordersByStand: AdminDashboardOrdersByStandItem[];
}

export interface UpdateStandStatusBody {
  status: StandStatus;
  reopenAt?: string;
  maxPendingOrders?: number;
}

export interface AssignRiderBody {
  riderId: string;
}

export type BroadcastBodyType =
  (typeof BroadcastBodyType)[keyof typeof BroadcastBodyType];

export const BroadcastBodyType = {
  promo: "promo",
  info: "info",
  halftime: "halftime",
} as const;

export interface BroadcastBody {
  message: string;
  section: string;
  type?: BroadcastBodyType;
}

export interface NotificationLog {
  id: string;
  stadiumId: string;
  message: string;
  section: string;
  type: string;
  sentAt: string;
}

export interface HeatmapCell {
  section: string;
  count: number;
}

export interface HalftimeBody {
  active: boolean;
  menuItemId?: string;
  discountPercent?: number;
  message?: string;
}

export interface HalftimePromo {
  stadiumId: string;
  active: boolean;
  menuItemId?: string;
  menuItemName?: string;
  standName?: string;
  imageUrl?: string;
  discountPercent?: number;
  message?: string;
  expiresAt?: string;
}

export interface TrendingItem {
  menuItemId: string;
  name: string;
  standId: string;
  standName: string;
  ordersLastHour: number;
  imageUrl: string;
  priceMad: number;
  section?: string;
}

export interface FeedEvent {
  id: string;
  message: string;
  section: string;
  standName: string;
  itemName: string;
  imageUrl?: string;
  createdAt: string;
}

export type LiveMatchStatus =
  (typeof LiveMatchStatus)[keyof typeof LiveMatchStatus];

export const LiveMatchStatus = {
  scheduled: "scheduled",
  first_half: "first_half",
  halftime: "halftime",
  second_half: "second_half",
  full_time: "full_time",
} as const;

export interface LiveMatch {
  stadiumId: string;
  homeTeam: string;
  awayTeam: string;
  homeFlag?: string;
  awayFlag?: string;
  homeScore: number;
  awayScore: number;
  minute: number;
  status: LiveMatchStatus;
  kickoffAt?: string;
}

export type ListStandsParams = {
  category?: string;
};

export type GetTrendingItemsParams = {
  section?: string;
};

export type GetAdminOrdersParams = {
  status?: string;
};