# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminLoginBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Admin" }
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /admin/{stadiumId}/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Admin KPIs and live overview for a stadium
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminDashboard" }
  /admin/{stadiumId}/orders:
    get:
      operationId: getAdminOrders
      tags: [admin]
      summary: Live order queue for the stadium
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
  /admin/{stadiumId}/riders:
    get:
      operationId: getAdminRiders
      tags: [admin]
      summary: All riders for the stadium with status
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Rider" }
  /admin/stands/{standId}/status:
    post:
      operationId: updateStandStatus
      tags: [admin]
      summary: Update stand status (open/temp_closed/closed)
      parameters:
        - name: standId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateStandStatusBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Stand" }
  /admin/orders/{orderId}/assign:
    post:
      operationId: assignOrderRider
      tags: [admin]
      summary: Assign or reassign a rider to an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssignRiderBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /admin/{stadiumId}/broadcast:
    post:
      operationId: broadcastNotification
      tags: [admin]
      summary: Broadcast a notification to fans in a stadium section
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BroadcastBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationLog" }
  /admin/{stadiumId}/broadcasts:
    get:
      operationId: listBroadcasts
      tags: [admin]
      summary: Recent broadcasts for the stadium
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NotificationLog" }
  /admin/{stadiumId}/heatmap:
    get:
      operationId: getOrderHeatmap
      tags: [admin]
      summary: Heatmap of orders by stadium section
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/HeatmapCell" }
  /admin/{stadiumId}/halftime:
    post:
      operationId: triggerHalftimeMode
      tags: [admin]
      summary: Trigger or end "Half-Time Rush" mode with a featured item
      parameters:
        - name: stadiumId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/HalftimeBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HalftimePromo" }
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
    SeatInfo:
      type: object
      required: [stadiumId, section, row, seat]
      properties:
        stadiumId: { type: string }
        section: { type: string }
        row: { type: string }
        seat: { type: string }
...
    StandStatus:
      type: string
      enum: [open, busy, temp_closed, closed]
    Stand:
      type: object
      required: [id, stadiumId, name, category, status, avgPrepTimeMinutes, pendingOrders, maxPendingOrders, waitTimeMinutes, imageUrl, rating]
      properties:
        id: { type: string }
        stadiumId: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        status: { $ref: "#/components/schemas/StandStatus" }
        avgPrepTimeMinutes: { type: integer }
        pendingOrders: { type: integer }
        maxPendingOrders: { type: integer }
        waitTimeMinutes: { type: integer }
        imageUrl: { type: string }
        rating: { type: number }
        location: { type: string }
        reopenAt: { type: string, format: date-time }
        isHalalOnly: { type: boolean }
...
    OrderItem:
      type: object
      required: [menuItemId, name, quantity, priceMad]
      properties:
        menuItemId: { type: string }
        name: { type: string }
        quantity: { type: integer }
        priceMad: { type: number }
        imageUrl: { type: string }
    OrderStatus:
      type: string
      enum: [received, preparing, on_the_way, delivered, cancelled]
    Order:
      type: object
      required: [id, sessionId, standId, standName, items, status, seat, subtotalMad, deliveryFeeMad, totalMad, etaMinutes, createdAt, isPreOrder]
      properties:
        id: { type: string }
        sessionId: { type: string }
        standId: { type: string }
        standName: { type: string }
        riderId: { type: string }
        riderName: { type: string }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        status: { $ref: "#/components/schemas/OrderStatus" }
        seat: { $ref: "#/components/schemas/SeatInfo" }
        subtotalMad: { type: number }
        deliveryFeeMad: { type: number }
        totalMad: { type: number }
        etaMinutes: { type: integer }
        createdAt: { type: string, format: date-time }
        scheduledFor: { type: string, format: date-time }
        deliveredAt: { type: string, format: date-time }
        rating: { type: integer }
        isPreOrder: { type: boolean }
        fanName: { type: string }
...
    Rider:
      type: object
      required: [id, name, stadiumId, status, activeOrders]
      properties:
        id: { type: string }
        name: { type: string }
        stadiumId: { type: string }
        status:
          type: string
          enum: [available, busy, offline]
        activeOrders: { type: integer }
        avatarUrl: { type: string }
...
    Admin:
      type: object
      required: [id, email, stadiumId, stadiumName]
      properties:
        id: { type: string }
        email: { type: string }
        stadiumId: { type: string }
        stadiumName: { type: string }
    AdminLoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AdminDashboard:
      type: object
      required: [ordersToday, revenueTodayMad, activeRiders, avgDeliveryMinutes, ordersByHour, topItems, ordersByStand]
      properties:
        ordersToday: { type: integer }
        revenueTodayMad: { type: number }
        activeRiders: { type: integer }
        avgDeliveryMinutes: { type: number }
        ordersByHour:
          type: array
          items:
            type: object
            required: [hour, count]
            properties:
              hour: { type: string }
              count: { type: integer }
        topItems:
          type: array
          items:
            type: object
            required: [name, count, standName]
            properties:
              name: { type: string }
              count: { type: integer }
              standName: { type: string }
        ordersByStand:
          type: array
          items:
            type: object
            required: [standName, count, revenueMad]
            properties:
              standName: { type: string }
              count: { type: integer }
              revenueMad: { type: number }
    UpdateStandStatusBody:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/StandStatus" }
        reopenAt: { type: string, format: date-time }
        maxPendingOrders: { type: integer }
    AssignRiderBody:
      type: object
      required: [riderId]
      properties:
        riderId: { type: string }
    BroadcastBody:
      type: object
      required: [message, section]
      properties:
        message: { type: string }
        section: { type: string }
        type:
          type: string
          enum: [promo, info, halftime]
    NotificationLog:
      type: object
      required: [id, stadiumId, message, section, type, sentAt]
      properties:
        id: { type: string }
        stadiumId: { type: string }
        message: { type: string }
        section: { type: string }
        type: { type: string }
        sentAt: { type: string, format: date-time }
    HeatmapCell:
      type: object
      required: [section, count]
      properties:
        section: { type: string }
        count: { type: integer }
    HalftimeBody:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
        menuItemId: { type: string }
        discountPercent: { type: integer }
        message: { type: string }
    HalftimePromo:
      type: object
      required: [stadiumId, active]
      properties:
        stadiumId: { type: string }
        active: { type: boolean }
        menuItemId: { type: string }
        menuItemName: { type: string }
        standName: { type: string }
        imageUrl: { type: string }
        discountPercent: { type: integer }
        message: { type: string }
        expiresAt: { type: string, format: date-time }
...