# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";
import type { SeatInfo } from "./seatInfo";

export interface Order {
  id: string;
  sessionId: string;
  standId: string;
  standName: string;
  riderId?: string;
  riderName?: string;
  items: OrderItem[];
  status: OrderStatus;
  seat: SeatInfo;
  subtotalMad: number;
  deliveryFeeMad: number;
  totalMad: number;
  etaMinutes: number;
  createdAt: Date;
  scheduledFor?: Date;
  deliveredAt?: Date;
  rating?: number;
  isPreOrder: boolean;
  fanName?: string;
}