# lib/api-zod/src/generated/types/standStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Stadium Eats API — in-stadium food ordering for FIFA World Cup Morocco 2030
 * OpenAPI spec version: 0.1.0
 */

export type StandStatus = (typeof StandStatus)[keyof typeof StandStatus];

export const StandStatus = {
  open: "open",
  busy: "busy",
  temp_closed: "temp_closed",
  closed: "closed",
} as const;